✨ feat:Prime datatable global function
