+++ server/apps/core/src/modules/account/account.service.ts
@@ -1,12 +1,13 @@
 import { AccountEditDTO } from '@core/account/dto/account.edit'
 import { AccountSignInDTO } from '@core/account/dto/account.signin'
 import { LogLogin, LogLoginDocument } from '@log/schemas/log.login'
-import { HttpStatus, Injectable } from '@nestjs/common'
+import { Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/mongoose'
 import { AuthService } from '@security/auth.service'
 import { GlobalResponse } from '@utility/dto/response'
 import { gen_uuid } from '@utility/generator'
 import { modCodes } from '@utility/modules'
+import { prime_datatable } from '@utility/prime'
 import { TimeManagement } from '@utility/time'
 import * as bcrypt from 'bcrypt'
 import { Model, Types } from 'mongoose'
@@ -29,103 +30,7 @@ export class AccountService {
   }
 
   async all(parameter: any) {
-    const first = parameter.first ? parseInt(parameter.first) : 0
-    const rows = parameter.rows ? parseInt(parameter.rows) : 20
-    const sortField = parameter.sortField ? parameter.sortField : 'created_at'
-    const sortOrder = parameter.sortOrder ? parseInt(parameter.sortOrder) : 1
-    const filters = parameter.filters
-    const query = []
-    const sort_set = {}
-
-    const filter_builder = { $and: [] }
-    const filterSet = filters
-    for (const a in filterSet) {
-      if (
-        a &&
-        a !== '' &&
-        filterSet[a].value !== '' &&
-        filterSet[a].value !== null
-      ) {
-        const autoColumn = {}
-        if (autoColumn[a] === undefined) {
-          autoColumn[a] = {}
-        }
-
-        if (filterSet[a].matchMode === 'contains') {
-          autoColumn[a] = {
-            $regex: new RegExp(`${filterSet[a].value}`, 'i'),
-          }
-        } else if (filterSet[a].matchMode === 'notContains') {
-          autoColumn[a] = {
-            $not: {
-              $regex: new RegExp(`${filterSet[a].value}`, 'i'),
-            },
-          }
-        } else if (filterSet[a].matchMode === 'endsWith') {
-          autoColumn[a] = {
-            $regex: new RegExp(`${filterSet[a].value}$`, 'i'),
-          }
-        } else if (filterSet[a].matchMode === 'equals') {
-          autoColumn[a] = {
-            $eq: filterSet[a].value,
-          }
-        } else if (filterSet[a].matchMode === 'notEquals') {
-          autoColumn[a] = {
-            $not: {
-              $eq: filterSet[a].value,
-            },
-          }
-        }
-
-        filter_builder.$and.push(autoColumn)
-      }
-    }
-
-    if (filter_builder.$and.length > 0) {
-      query.push({
-        $match: filter_builder,
-      })
-    } else {
-      query.push({
-        $match: {
-          $and: [{ deleted_at: null }],
-        },
-      })
-    }
-    //---------------------------------------------------------------------------
-
-    const allNoFilter = await this.accountModel.aggregate(
-      query,
-      (err, result) => {
-        return result
-      }
-    )
-
-    query.push({ $skip: first })
-
-    query.push({ $limit: rows })
-
-    if (sortField && sortOrder) {
-      if (sort_set[sortField] === undefined) {
-        sort_set[sortField] = sortOrder
-      }
-
-      query.push({
-        $sort: sort_set,
-      })
-    }
-
-    const data = await this.accountModel.aggregate(query, (err, result) => {
-      return result
-    })
-
-    return {
-      message: HttpStatus.OK,
-      payload: {
-        totalRecords: allNoFilter.length,
-        data: data,
-      },
-    }
+    return await prime_datatable(parameter, this.accountModel)
   }
 
   async detail(_id: string): Promise<AccountModel> {
